+++ modules/LifeGame/LifeGameMapBase.hh
@@ -18,9 +18,11 @@ class LifeGameMapBase : public WindowBase {
 
     virtual void UpdateData() = 0;
 
+    virtual void UpdateSize(int width, int history) = 0;
+
     virtual void RandomElems(int rate = 50) = 0;
 
-    virtual void UpdateMap(LifeGameRuleBase& rule) = 0;
+    virtual void UpdateMap(LifeGameRuleBase& rule, int rule_id) = 0;
 };
 
 #endif