+++ modules/LifeGame/LifeGameMapOne.hh
@@ -7,12 +7,16 @@
 
 class LifeGameMapOne : public LifeGameMapBase {
   public:
-    size_t        size;
-    LifeGameElem* map;
+    float          bsize;
+    float          padding;
+    size_t         size;
+    size_t         history;
+    size_t         index;
+    LifeGameElem** map;
 
     LifeGameMapOne();
 
-    LifeGameMapOne(size_t size_, bool ring_);
+    LifeGameMapOne(size_t size_, size_t his_);
 
     ~LifeGameMapOne();
 
@@ -20,9 +24,11 @@ class LifeGameMapOne : public LifeGameMapBase {
 
     virtual void UpdateData();
 
+    virtual void UpdateSize(int width, int history);
+
     virtual void RandomElems(int rate = 50);
 
-    virtual void UpdateMap(LifeGameRuleBase& rule);
+    virtual void UpdateMap(LifeGameRuleBase& rule, int rule_id);
 };
 
 #endif